+++ src/main/java/edu/online/messenger/controller/UserController.java
@@ -1,0 +1,21 @@
+package edu.online.messenger.controller;
+
+import edu.online.messenger.service.UserService;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
+import lombok.AllArgsConstructor;
+
+@RestController
+@RequestMapping("/api/users")
+@AllArgsConstructor
+public class UserController {
+
+    private final UserService userService;
+
+    @GetMapping("existence/id/{userId}")
+    public boolean existsById(@PathVariable Long userId) {
+        return userService.existsById(userId);
+    }
+}